Scarlet Death + Victory anim, God Mode (press "0" to toggle), Health Potion Loading, Longer Death Screen
and a couple of misc. bugfixes.
(I should really make more commits...)

+++ unity/Scarlet_Production/Assets/Scripts/Player/PlayerHealth.cs
@@ -6,7 +6,7 @@ public class PlayerHealth : CharacterHealth {
 
     protected override void Start()
     {
-        m_CurrentHealth = PlayerPrefs.GetInt("MaxHP", (int) m_MaxHealth);
+        m_CurrentHealth = PlayerPrefs.GetFloat("MaxHP", m_MaxHealth);
         m_MaxHealth = m_CurrentHealth;
         m_HealthStart = m_CurrentHealth;
         m_HealthOld = m_CurrentHealth;
